requirements.txt
altgraph==0.17.3
pefile==2023.2.7
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyserial==3.5
pywin32-ctypes==0.2.0
websockets==10.4
